Refactoring and other small updates
