+++ src/components/Layout.jsx
@@ -1,5 +1,5 @@
-import Header from './Header';
-import Footer from './Footer';
+import Header from './ui/Header';
+import Footer from './ui/Footer';
 import { Outlet } from 'react-router-dom';
 
 const Layout = () => {
