+++ src/components/ui/Header.jsx
@@ -1,8 +1,8 @@
 import { useState, useContext } from 'react';
 import { FaBars, FaCartShopping, FaX, FaStore } from "react-icons/fa6";
-import NavList from './NavList';
+import NavList from '../NavList';
 import { Link } from 'react-router-dom';
-import CartContext from '../services/providers/CartContext';
+import CartContext from '../../services/providers/CartContext';
 
 const Header = () => {
   const [navbar, setNavbar] = useState(false);
