+++ src/components/ui/Modal.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Modal = () => {
+  return (
+    <div>
+      
+    </div>
+  )
+};
+
+export default Modal;
