+++ src/pages/Product/Product.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import Loader from '../../components/Loader';
+import Loader from '../../components/ui/Loader';
 import SingleProduct from '../../components/SingleProduct';
 import { useApi } from '../../hooks/useApi';
